Reorganise instance resource
In this PR we:
- split instance.go into
  - resource.go
  - instance_read.go
  - instance_create.go
  - instance_delete.go
  - instance_import.go
- we rename update.go to instance_update.go
- we run gofumpt and goimports in the instance package
