+++ internal/framework/subproviders/morpheus/resources/instance/instance_import.go
@@ -1,0 +1,30 @@
+package instance
+
+import (
+	"context"
+	"strconv"
+
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+// ImportState implements resource.ResourceWithImportState.
+func (g *Resource) ImportState(
+	ctx context.Context,
+	req resource.ImportStateRequest,
+	resp *resource.ImportStateResponse,
+) {
+	id, err := strconv.Atoi(req.ID)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"import instance resource",
+			"provided import ID '"+req.ID+"' is invalid (non-number)",
+		)
+
+		return
+	}
+
+	diags := resp.State.SetAttribute(ctx, path.Root("id"), id)
+
+	resp.Diagnostics.Append(diags...)
+}
