+++ internal/framework/subproviders/morpheus/resources/instance/resource.go
@@ -1,0 +1,60 @@
+package instance
+
+import (
+	"context"
+	"errors"
+	"slices"
+	"strings"
+
+	"github.com/cenkalti/backoff/v5"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/configure"
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/constants"
+)
+
+var (
+	_ resource.Resource                = &Resource{}
+	_ resource.ResourceWithImportState = &Resource{}
+)
+
+func NewResource() resource.Resource {
+	return &Resource{}
+}
+
+type Resource struct {
+	configure.ResourceWithMorpheusConfigure
+}
+
+// Metadata implements resource.Resource.
+func (g *Resource) Metadata(
+	_ context.Context,
+	req resource.MetadataRequest,
+	resp *resource.MetadataResponse,
+) {
+	resp.TypeName = req.ProviderTypeName + "_" + constants.SubProviderName + "_instance"
+	resp.TypeName = strings.Join(
+		[]string{req.ProviderTypeName, constants.SubProviderName, "instance"},
+		"_",
+	)
+}
+
+// Schema implements resource.Resource.
+func (g *Resource) Schema(
+	ctx context.Context,
+	_ resource.SchemaRequest,
+	resp *resource.SchemaResponse,
+) {
+	resp.Schema = InstanceResourceSchema(ctx)
+}
+
+func checkStatusDone(status string, targetStatuses []string, errorStatuses []string) error {
+	switch {
+	case slices.Contains(errorStatuses, status):
+		return backoff.Permanent(errors.New("reached error status: " + status))
+	case slices.Contains(targetStatuses, status):
+		return nil
+	default:
+		return backoff.RetryAfter(5)
+	}
+}
